event_mgmt_platform/backend/requirements.txt
APScheduler==3.11.0
Django==5.1.7
django_apscheduler==0.7.0
# djangorestframework==3.14.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn
django-cors-headers
psycopg2-binary